ip: improve endianness adjustment
Rather than hacking around the inability to bind references to unaligned
scalars, wrap them in a struct (which can then be marked as unaligned),
and use the wrapped values everywhere.  This ensures the code works on
platforms that need special instructions for unaligned access.

Adjust the forwarding functions to require references; passing by value
accidentally can lose the conversion.