net: tcp
Rudimentary TCP support (establishes a connection then falls over, lots of
things missing).

In a departure from tradition, we don't have a state enum with all the
traditional LISTENING states etc.  Instead we have boolean values
indicating whether an event (like a remote SYN received, or our SYN
acknowledged) has happened.  This simplifies things, because in TCP
the sending side and the receiving side are mostly orthogonal, while the
TCP state mixes the two.

TCP is implemented as a template, in order to tune it for the differences
in TCP/IPv4 vs. TCP/IPv6 (mostly address size and pseudo header).