Instructions on how to run Seastar on OSv
Due to popular demand, this patch adds README-OSv, containing detailed
instructions on how to run Seastar inside OSv.

Because Seastar is moving at a fast pace, and uses new "exotic" Linux
features almost every day, please make sure that you use the latest
master of OSv, otherwise you are very likely to be missing features that
Seastar needs. At the time of this writing, you'll even need to apply a
few uncommited OSv patches (those which add thread pinning support), but
hopefully these patches will be committed today and the OSv master will
go back to enough for the Seastar master.

To clarify (though this is also mentioned in the README-OSv itself),
Seastar can only run in OSv with the "posix" network stack. The
"native" network stack does not work yet, because we haven't completed
a mechanism to assign virtio (or xen) from OSv directly to Seastar.
This is in the works.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>