virtio: don't use file eventfd for OSv notifications
Now that our reactor supports non-file-descriptor notification
mechanisms, switch to using one instead of eventfd when notifying
of virtio interrupts.

This will allow us to change the OSv enable_interrupt() code to
run the handler directly, not in a separate thread, because it
no longer needs to do sleepable write() to an eventfd file descriptor.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>