xen: use the right index
The index in the ring and the packet id tends to be the same. But it doesn't
have to.  There are some situations where the backend and the frontend get out
of sync with this, and this is totally valid.

One example is when the backend skb already have enough room to hold all of the
data being transmitted (netback.c, line 1611 @3.16). The netback will respond
immediately, even though there are other pending packets that are not yet fully
processed.

The ring index, then, must come from the rsp value, not from the req/rsp id.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>